Move files to database folder
