+++ core/classes/bot-settings.js
@@ -1,5 +1,5 @@
 const BotSettingsModel = require("./../../shared-models/bot-settings.js");
-const db = require("./../apis/database-crud.js");
+const db = require("./../database/crud.js");
 const Logger = require("../utils/logger.js");
 const discordClientProvider = require("./../apis/discord-client-provider.js");
