+++ core/classes/user-role.js
@@ -1,5 +1,5 @@
 const UserRoleModel = require("./../../shared-models/user-role.js");
-const db = require("./../apis/database-crud.js");
+const db = require("./../database/crud.js");
 const Logger = require("../utils/logger.js");
 const discordClientProvider = require("./../apis/discord-client-provider.js");
 /**
